Add Makefile for build, run, and test commands
A new Makefile has been introduced to streamline the processes of building, running, and testing. It includes the 'build' command to compile the source code, 'run' to first build the code and then execute the output program, and 'test' to run the unit tests with verbosity enabled. This aims to simplify common tasks and improve development workflow.